requirements.txt
Flask
flask-cors
gunicorn
python-dotenv
langchain
langchain-community
langchain-google-genai
langchain-huggingface
sentence-transformers
chromadb
unstructured[xlsx]
PyMuPDF
pandas
numpy
scikit-learn
xgboost
matplotlib
seaborn
joblib